Split the account-email route into shared helpers and a detail panel.
account-email.tsx was over the client-route budget. Loader, query helpers,
and list-cell styles move to account-email-shared.ts; the selected-message
panel moves to account-email-detail.tsx. The route keeps load, classify,
and the inbox table.
